Move(#61) CountIndicator 컴포넌트 이동
- common/user-page -> common
